requirements.txt
google-cloud-aiplatform
python-docx